refactor Jobs and Logs api folder structure
