+++ src/api/index.ts
@@ -1,8 +1,35 @@
 import {
+  cancelJob,
+  cancelJobs,
+  deleteJob,
+  deleteJobs,
+  executeJob,
+  executeJobQuery,
+  fetchJobCount,
+  fetchJobs,
+  fetchLastJob,
+} from './Jobs/JobsApi';
+import {
   deleteJsonDocument,
   fetchJsonDocument,
   fetchJsonDocuments,
   postJsonDocuments,
 } from './JsonDocuments/JsonDocumentsApi';
+import { fetchLogs } from './Logs/LogsApi';
 
-export { fetchJsonDocuments, fetchJsonDocument, postJsonDocuments, deleteJsonDocument };
+export {
+  fetchJsonDocuments,
+  fetchJsonDocument,
+  postJsonDocuments,
+  deleteJsonDocument,
+  executeJobQuery,
+  executeJob,
+  cancelJob,
+  cancelJobs,
+  fetchJobs,
+  deleteJob,
+  deleteJobs,
+  fetchLastJob,
+  fetchJobCount,
+  fetchLogs,
+};
